cond: add counters for sequence
It basically worked. However, if a thread is calling pi_cond_wait() and
is interrupted after dropping the mutex then it could happen that the
syscall is entered _after_ the wakeup occured which it missed. So here
we have counters which are modified untder the private lock:
- The futex (cond) id is incremented during each wait / wakeup
- If the waiter raced against a waker then the syscall will return
  EAGAIN which means he has to retry. If wake_id >= larger than the
  futex value then it means that is was woken up but missed it. In that
  case it is save to return as woken up. Otherwise the operation has to
  be retried.

The IDs magic seems to work and is not overlfow safe. Also the timeout
part is missing

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>