pi_cond: Add sequence counters
It basically worked. However, if a thread is calling pi_cond_wait() and
is interrupted after dropping the mutex then it could happen that the
syscall is entered _after_ the wakeup occured which it missed. So here
we have counters which are modified untder the private lock:

The futex (cond) id is incremented during each wait and wakeup.

If the waiter raced against a waker then the syscall will return EAGAIN
and the waiter has to retry. If wake_id >= the futex value then it was
woken but missed it. In that case it is safe to return as successfully
woken. Otherwise the operation has to be retried.

Notes / TODO:
 - The id magic is not overlfow safe.
 - Timeouts are not implemented

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
[dvhart: use cond->mutex as target for futex_wait_requeue_pi() calls]
Signed-off-by: Darren Hart (VMware) <dvhart@infradead.org>