pi_cond, pi_mutex: Add support for static definition and initialization
A fair amount of existing code which uses pthread_mutex* and
pthread_cond* makes use of static definitions/initialization for
mutexes and conditional variables. Providing a way to statically
define and initialize the pi_mutex* and pi_cond* equivalents in
librtpi would be beneficial when porting existing code.

Additionally and more importantly having a way to statically define a
pi mutex or conditional variable avoids dynamic memory allocations in
an RT thread which can lead to latency increases.

Provide DEFINE_PI_* macros modeled after the Linux kernel's
"DEFINE_MUTEX" APIs for statically initializing a pi_mutex_t and/or
pi_cond_t variable.

One downside of providing a way to statically initialize the pi mutex
and condvar is the fact that the corresponding structure sizes become
part of the librtpi ABI. In order to mitigate this and provide space
for future enhancements make use of a union in combination with a
anonymous struct and a pad field to set the pi_mutex size to 64 bytes
and pi_cond size to 128 bytes. For pi_mutex this matches the common L1
cache line size on 64-bit machines and in combination with setting the
alignment provides performance advantages in addition to extra space
for future enhancements. Same reasoning applies to the pi_cond data
type except it spans two cache lines with the first 64 bytes being
occupied by the embedded private mutex. This should allow for some
runway before the ABI version for librtpi needs to change.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>