pi_cond, pi_mutex: Add support for static definition and initialization
A fair amount of existing code which uses pthread_mutex* and
pthread_cond* makes use of static definitions/initialization for
mutexes and conditional variables. Providing a way to statically
define and initialize the pi_mutex* and pi_cond* equivalents in
librtpi would be beneficial when porting existing code.

Additionally and more importantly having a way to statically define a
pi mutex or conditional variable avoids dynamic memory allocations in
an RT thread which can lead to latency increases.

Provide DEFINE_PI_* macros modeled after the Linux kernel's
"DEFINE_MUTEX" APIs for statically initializing a pi_mutex_t and/or
pi_cond_t variable.

One downside of providing a way to statically initialize the pi mutex
and condvar is the fact that the corresponding structure sizes become
part of the librtpi ABI. In order to mitigate this and provide space
for future enhancements make use of a union in combination with a
anonymous struct and a pad field to set the size to the WAG estimate
of roughly double the current structure size. This should allow for
some runway before the ABI version for librtpi needs to change.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>