glibc-tests/tst-cond22: Remove prints of condvar internals
Imported glibc test 'tst-cond22' prints the internal fields of the
conditional variable data structure in a couple of places for debug
purposes. The prints have no effect on the outcome of the test.

librtpi has a different layout for the condvar data structure and the
internal fields of this structure shouldn't be accessed directly.
Remove these debug prints.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>