pi_mutex, pi_cond: Return a positive error number on failure
Per POSIX mutex and condvar operations are expected to return 0 on
success and a positive error number on failure.

The current implementation uses the futex syscall() return value
directly in several places. The syscall() call returns 0 on success
and -1 to indicate an error with the error code stored in 'errno'
variable.

Change the code to return the 'errno' value when the futex syscall
fails. This matches the expected POSIX behavior and fixes:
tests/glibc-tests/tst-cond9.c

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>