pi_mutex: Return early from pi_mutex_lock() on deadlock
When locking a mutex if the fast path call to pi_mutex_trylock() fails
with EDEADLK because the current thread already owns the mutex there
is no point in doing a futex syscall just to get the same error. Avoid
the additional syscall and return early with appropriate error code.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>