pi_cond: Re-acquire the user mutex on error
On error the kernel does not lock the user mutex for us but per POSIX[1]
condvar wait operations are expected to always return with the user mutex
locked. Re-acquire the user mutex in the error path before returning.

Relevant POSIX section:

"Even though the dynamic binding between condition variable and mutex may
be removed or replaced between the time a thread is unblocked from a wait
on the condition variable and the time that it returns to the caller or
begins cancellation cleanup, the unblocked thread shall always re-acquire
the mutex specified in the condition wait operation call from which it is
returning."
[1] https://pubs.opengroup.org/onlinepubs/9699919799/functions/pthread_cond_timedwait.html

Fixes: tests/glibc-tests/tst-cond5.c
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>