glibc-tests: Enable tst-cond5.c
tests/glibc-tests/tst-cond5.c tests correct operation of
pi_cond_timedwait() in a timeout scenario.

Prior to commit 9835bf5b6f9c ("pi_cond: Re-acquire the user mutex on
error") this test was failing on the pi_mutex_unlock() call after the
pi_cond_timedwait() call times out. The failure was caused by the fact that
in a timeout scenario the kernel will not re-acquire the user mutex for us
and the mutex was already dropped before calling into the kernel. This in
effect results in a double unlock scenario for the user mutex on error.

The commit mentioned above fixed this scenario by re-acquiring the user
mutex on error before returning from pi_cond_timedwait(). We can re-enable
tst-cond5.c as a result.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>