rtpi: add a .clang-format
This configuration is just copied from Linux, and it's close enough
to matching the code style of the rtpi code already present (though
not the test code that came via glibc).

It's in the "rtpi" subdir to only apply to the new code in the C++
wrappers instead of applying to other code under `src`. No reformatting
on existing files is being done at this time.

Signed-off-by: Brenda Streiff <brenda.streiff@ni.com>