tests: libstdc++-tests: add rtpi-ized versions of libstdc++ tests
This is a scaled-down version of the libstdc++ test suite, that only
includes the mutex and condition_variable tests, and has been converted
over to use the librtpi C++ wrappers.

This is also restricted to the set of tests that compile successfully;
libstdc++ includes some tests that intentionally fail compilation
(in order to make sure that deleted methods are truly deleted) and uses
a bunch of scripts (some in dejagnu) to pick out those errors, and I
opted to not try to port that over.

testsuite_common_types.h is also scaled down to remove a lot of the
integer-type-list helpers (which the tests we have don't need), because
they don't compile successfully with the g++ 5.x that's currently used
in CI.

Signed-off-by: Brenda Streiff <brenda.streiff@ni.com>